MET #20: First working version of two model comparison app after refactor.
